auth-service/pyproject.toml
[project]
name = "auth-service"
version = "0.23.0"
requires-python = ">=3.11"
dependencies = ["fastapi", "uvicorn[standard]", "sqlalchemy>=2.0.0", "aiomysql", "python-jose[cryptography]", "passlib[bcrypt]", "python-dotenv", "pydantic", "cryptography", "python-multipart", "email-validator", "bcrypt==4.0.1"]
